+++ src/Blog/templates/Home/HomeTemplate.component.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { Head } from 'react-static';
 
-import Link from '../../../Shared/elements/Link/Link.component';
-import Markdown from '../../../Shared/elements/Markdown/Markdown.component';
-import DefaultLayout from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+import { Link } from '../../../Shared/elements/Link/Link.component';
+import { Markdown } from '../../../Shared/elements/Markdown/Markdown.component';
+import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 import { ContentPage } from '../../../Shared/types/Page.model';
 
 import { Post } from '../../types/Post.model';