+++ src/Pages/templates/About/AboutTemplate.component.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Head } from 'react-static';
 
-import Markdown from '../../../Shared/elements/Markdown/Markdown.component';
-import DefaultLayout from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+import { Markdown } from '../../../Shared/elements/Markdown/Markdown.component';
+import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 import { ContentPage } from '../../../Shared/types/Page.model';
 
 interface Props {