+++ src/Shared/lib/classes/TemplateLocator.ts
@@ -1,8 +1,6 @@
-class TemplateLocator {
+export class TemplateLocator {
   public locate(name: string): string {
     const [module, container] = name.split('/');
     return `src/${module}/containers/${container}/${container}Container.component`;
   }
 }
-
-export default TemplateLocator;