+++ src/Site/containers/NotFound/NotFoundContainer.component.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { useRouteData } from 'react-static';
 
 import { bodyClassSwitch } from '../../../Shared/lib/dom';