+++ src/routes.ts
@@ -1,15 +1,15 @@
-import TemplateLocator from './Shared/lib/classes/TemplateLocator';
+import { TemplateLocator } from './Shared/lib/classes/TemplateLocator';
 
-import loadPosts from './Blog/posts.source';
-import loadDocs from './Docs/docs.source';
-import loadPages from './Pages/pages.source';
+import { loadPosts } from './Blog/posts.source';
+import { loadDocs } from './Docs/docs.source';
+import { loadPages } from './Pages/pages.source';
 
-import buildBlogRoutes from './Blog/blog.routes';
-import buildDocsRoutes from './Docs/docs.routes';
-import buildPagesRoutes from './Pages/pages.routes';
-import buildSiteRoutes from './Site/site.routes';
+import { buildRoutes as buildBlogRoutes } from './Blog/blog.routes';
+import { buildRoutes as buildDocsRoutes } from './Docs/docs.routes';
+import { buildRoutes as buildPagesRoutes } from './Pages/pages.routes';
+import { buildRoutes as buildSiteRoutes } from './Site/site.routes';
 
-const routeBuilder = () => {
+export const routeBuilder = () => {
   const templates = new TemplateLocator();
 
   const getRoutes = async () => {
@@ -27,5 +27,3 @@ const routeBuilder = () => {
 
   return { getRoutes };
 };
-
-export default routeBuilder;